# storage/api-docs/api-docs.json
{"swagger":"2.0","info":{"title":"OpenApi","contact":{"name":"Pek Ratanak","url":"https:\/\/www.google.com"},"version":"3.0"},"host":"homestead.test","basePath":"","schemes":["http","https"],"paths":{"\/api\/test":{"get":{"tags":["\u79d8\u94a5\u751f\u6210"],"summary":"Get user","description":"\u8fd4\u56de\u6d4b\u8bd5\u5185\u5bb9","operationId":"api.dashboard.index","produces":["application\/json"],"parameters":[{"name":"reason","in":"query","description":"\u62ff\u6570\u636e\u7684\u7406\u7531","required":true,"type":"string"}],"responses":{"200":{"description":"Dashboard overview."},"401":{"description":"Unauthorized action."}},"security":[{"passport":[]}]}},"\/api\/ip":{"get":{"tags":["IP\u8303\u56f4"],"summary":"IP","description":"\u8fd4\u56deip\u8303\u56f4","operationId":"api.dashboard.index","produces":["application\/json"],"parameters":[{"name":"ip1","in":"query","description":"\u7b2c\u4e00\u4e2aIP\u5730\u5740","required":true,"type":"string"},{"name":"ip2","in":"query","description":"\u7b2c\u4e8c\u4e2aIP\u5730\u5740","required":true,"type":"string"}],"responses":{"200":{"description":"Dashboard overview."},"401":{"description":"Unauthorized action."}}}},"\/api\/norepeat":{"post":{"tags":["\u6570\u7ec4"],"summary":"Array","description":"\u53bb\u6570\u7ec4\u4e2d\u4e0d\u91cd\u590d\u7684\u6570","operationId":"api.dashboard.index","produces":["application\/json"],"parameters":[{"name":"data","in":"body","description":"\u6570\u7ec4","required":true,"schema":{"$ref":"#"}}],"responses":{"200":{"description":"Dashboard overview."}}}},"\/api\/json":{"get":{"tags":["\u8fd4\u56de\u6570\u636e"],"summary":"Array","description":"\u6d4b\u8bd5\u4f7f\u7528\u7684\u63a5\u53e3","operationId":"api.dashboard.index","produces":["application\/json"],"parameters":[{"name":"data","in":"body","description":"\u6570\u7ec4","required":true,"schema":{"$ref":"#"}}],"responses":{"200":{"description":"\u8fd4\u56de\u6210\u529f"}}}},"\/api\/vue-element-admin\/user\/login":{"post":{"tags":["Vue-api"],"summary":"Login","description":"Vue \u767b\u9646","operationId":"api.dashboard.index","produces":["application\/json"],"parameters":[{"name":"username","in":"query","description":"\u7528\u6237\u540d","required":true,"type":"string"},{"name":"password","in":"query","description":"\u5bc6\u7801","required":true,"type":"string"}],"responses":{"200":{"description":"Dashboard overview."},"401":{"description":"Unauthorized action."}},"security":[{"passport":[]}]}}},"definitions":{"Error":{"required":["code","message"],"properties":{"code":{"description":"Class Controller","type":"integer","format":"int32"},"message":{"description":"Class Controller","type":"string"}}}},"securityDefinitions":{"passport":{"type":"oauth2","description":"Laravel passport oauth2 security.","flow":"password","tokenUrl":"http:\/\/localhost\/oauth\/token","scopes":[]}}}